+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -121,7 +121,11 @@ class GeckoEngineTest {
                 TrackingProtectionPolicy.CONTENT,
                 TrackingProtectionPolicy.TEST,
                 TrackingProtectionPolicy.CRYPTOMINING,
-                TrackingProtectionPolicy.FINGERPRINTING
+                TrackingProtectionPolicy.FINGERPRINTING,
+                TrackingProtectionPolicy.SAFE_BROWSING_HARMFUL,
+                TrackingProtectionPolicy.SAFE_BROWSING_UNWANTED,
+                TrackingProtectionPolicy.SAFE_BROWSING_MALWARE,
+                TrackingProtectionPolicy.SAFE_BROWSING_PHISHING
         ).categories, contentBlockingSettings.categories)
         assertEquals(defaultSettings.trackingProtectionPolicy, TrackingProtectionPolicy.all())
 
@@ -174,7 +178,11 @@ class GeckoEngineTest {
             TrackingProtectionPolicy.CONTENT,
             TrackingProtectionPolicy.TEST,
             TrackingProtectionPolicy.CRYPTOMINING,
-            TrackingProtectionPolicy.FINGERPRINTING
+            TrackingProtectionPolicy.FINGERPRINTING,
+            TrackingProtectionPolicy.SAFE_BROWSING_HARMFUL,
+            TrackingProtectionPolicy.SAFE_BROWSING_UNWANTED,
+            TrackingProtectionPolicy.SAFE_BROWSING_MALWARE,
+            TrackingProtectionPolicy.SAFE_BROWSING_PHISHING
         ).categories, contentBlockingSettings.categories)
         assertTrue(engine.settings.testingModeEnabled)
         assertEquals("test-ua", engine.settings.userAgentString)